python-service/requirements.txt
httpx>=0.25.0
pika>=1.3.2
python-dotenv>=1.0.0
schedule>=1.2.0

